+++ spec/unit/virtus/class_methods/inherited_spec.rb
@@ -5,15 +5,15 @@ describe Virtus::ClassMethods, '#inherited' do
 
   let(:object) { Class.new { extend Virtus::ClassMethods } }
 
-  it 'includes an AttributesAccessor module' do
+  it 'includes an AttributeSet module' do
     descendant = subject
 
     # return the descendant's attribute accessor modules (superclass + own)
-    modules = descendant.ancestors.grep(Virtus::AttributesAccessor)
+    modules = descendant.ancestors.grep(Virtus::AttributeSet)
     modules.size.should be(2)
 
     # remove the superclass' attribute accessor module
-    modules -= object.ancestors.grep(Virtus::AttributesAccessor)
+    modules -= object.ancestors.grep(Virtus::AttributeSet)
 
     # the descendant should have it's own attribute accessor module
     modules.size.should be(1)
